+++ vmm/src/lib.rs
@@ -30,6 +30,7 @@ pub mod api;
 pub mod config;
 pub mod cpu;
 pub mod device_manager;
+pub mod memory_manager;
 pub mod vm;
 
 #[cfg(feature = "acpi")]
